feat(nav): consolidate account/earn pages into Portfolio + Earn tabbed hubs
Perp-DEX-standard IA (Hyperliquid/dYdX/Drift all use one Portfolio hub). Cuts
the scattered "your stuff" pages down to two tabbed hubs:

- /portfolio hub — tabs: Overview (dashboard), Positions, Wallet, My Markets.
- /earn hub — tabs: LP Vault, Stake.

Each source page is composed as a tab (only the active tab mounts, so no extra
data hooks). Current /portfolio (positions) extracted to
components/portfolio/PortfolioPositionsView; current /earn extracted to
components/earn/EarnVaultView. Tab bar matches the site (accent underline on
active); tabs are hash-deep-linkable (/portfolio#wallet, /earn#stake).

Header nav slimmed: Trade group = Markets/Portfolio/Earn; Build group =
Create/Developers/Faucet (Dashboard/Wallet/My Markets/Stake are now tabs). Old
routes (/dashboard /wallet /my-markets /stake) stay working standalone.

Portfolio.test repointed to the extracted PortfolioPositionsView.
